Move square related extension to Square
