+++ core/src/main/scala/Piece.scala
@@ -3,8 +3,6 @@ package chess
 import cats.Eq
 import cats.derived.*
 
-import Bitboard.*
-
 case class Piece(color: Color, role: Role) derives Eq:
 
   def is(c: Color)   = c == color
