+++ core/src/main/scala/format/pgn/Dumper.scala
@@ -5,7 +5,6 @@ object Dumper:
 
   def apply(board: Position, data: chess.Move, next: Position): SanStr =
     import data.*
-    import Bitboard.*
 
     val base = (promotion, piece.role) match
       case _ if castles =>
